Move tests for TranspilerRepository into their own corresponding test suite.
